Move aliases.js -> util subdir. Clean up and simplify path determination
